Node.js working with MongoDB
This is all "hardwired" just to demonstrate a little of what the interaction with the back-end database looks like on the server side.  NOT a full implementation at all, we are just finding restaurants in a certain zipcode area and building the JSON for that data! :)   Note the importance of good plumbing!